server/recommend/requirements.txt
openai
python-dotenv
pymongo
opensearch-py
requests-aws4auth
boto3
ijson
numpy
chromadb
